requirements.txt
arrow==1.3.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.6
jwcrypto==1.5.3
oauthlib==3.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauth==0.4.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
withings-api==2.4.0
